Makes _ optional in =let* and reindents code.
(=let* (((foo))
       (var (bar))
       ((baz)))
  ...)

is now a supported synonym for:

(=let* ((_ (foo))
       (var (bar))
       (_ (baz)))
  ...)
This change makes =let* function more like Haskell's do-notation.

The code is reindented using Emacs's default settings (Note:
indent-tabs-mode was nil, the tab width was 4 spaces, and Slime was
loaded along with smug).

The documentation is updated to reflect the changes in this commit.
Also, an improperly-closed SRC block is now closed.

Spurious blank lines at the end of source files are removed.