Make the restart work correctly.
When the restart is active, it can be triggered via the
(replace-invalid) function.  For this to work, the parser that throws
the error cannot be being run directly, but must be called by another
parser.  When the restart is triggered, it unwinds the parser stack one
parser higher than the parser that signaled the condition and then
fixes its input and reruns it.